feat: 권한 검증 로직 분리 #122
